pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto_test"
version = "0.1.0"
description = "My awesome package"
readme = "README.md"
authors = [{name = "chao", email = "178754558@qq.com"}]
classifiers = [
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
    # 核心依赖
    "playwright>=1.40.0",
    "allure-python-commons>=2.13.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "pydantic",
    "allure-pytest",
    
    # 数据库相关依赖
    "sqlalchemy>=2.0",
    "pymysql",
    "psycopg2-binary",
    "tqdm",
    "faker",

    # 用于环境检测
    "psutil>=5.9.0",  

    # 其他
    "opencv-python", # 额外功能（如 SIFT、SURF 等），可安装扩展包 pip install opencv-contrib-python
    "numpy",
    "cryptography",
    "scikit-image"
    
]

[project.optional-dependencies]
dev = [
]

[tool.setuptools.packages.find]
where = ["src"]

